tietokantataulun aika muutettu integereiksi

+++ application/tasks/forms.py
@@ -1,13 +1,12 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, DecimalField, SelectMultipleField, validators
+from wtforms import StringField, IntegerField, validators
 from application.tags.models import Tag
 
 class TaskForm(FlaskForm):
     name = StringField('Task name', [validators.Length(min=2, max=144)])
     description = StringField('Description', [validators.Length( max=1000)])
-    estimatedtime = DecimalField('Estimated time', [validators.NumberRange(min=1, max=999)])
-    usedtime = DecimalField('Time used')
-#    tags = SelectMultipleField(u'Tags') #, coerce=int)
+    estimatedtime = IntegerField('Estimated time', [validators.NumberRange(min=1, max=999)])
+    usedtime = IntegerField('Time used')
  
     class Meta:
         csrf = False
